+++ app/account/(modalform)/page.tsx
@@ -1,0 +1,2 @@
+// We need this page file to use the common layout.
+export default function FakeAccountPage() {}
